Add new data set class for the fabrication data
A new data set class for the fabrication data is added. The class can select frames from specific video in order to split the data into train and validation data and it can sub-sample the frames to speed up training by avoiding the repetition of similar frames in the training data set.