Compute confusion matrices for fabrication data set
This file generates the confusion matrices for the cross validation of the fabrication data set. The keys in the dictionary are the index of the video used as test data and the value is the check point used for the model. The check point should be the result of a training that did not included the video of the given index.