Add data augmentation and support for custom data sets
The data is now augmented with a random 10 deg rotation, a random horizontal and vertical flip and a some color jitter to improve generalization of the model to new data. Furthermore, the datasets now use custom labels from which bad example where remmoved and to which relevant examples for the circular saw were added. It is now also possible to only use a subset of labels.