Confirm cpp and python input transforms are consistent
The output of the cpp inference and the python inference was different. This turned out to be the case be cause the pre-processing of the images were not consistent. The code of this commit can be used to verify that the inputs are the same. We found that the images have to be cropped into a square before they are resized.